backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.6.0
pymongo>=4.10.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
PyPDF2>=3.0.1
anthropic>=0.42.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
jinja2>=3.1.4
httpx>=0.28.0
python-dateutil>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
openpyxl>=3.1.0
